requirements.txt
pandas>=2.2.1
tqdm>=4.66.2
pandas_market_calendars>=4.4.0
python-dotenv>=1.0.1
mmengine>=0.10.3
distributed
einops>=0.7.0
timm>=0.9.16
iopath>=0.1.10
scipy>=1.12.0
tensorboard>=2.16.2
scikit-learn>=1.4.1
pyecharts>=2.0.5
snapshot_selenium>=0.0.2
pandas_ta>=0.3.14b0
colossalai>=0.3.7
accelerate>=0.29.3
diffusers>=0.27.2
ftfy>=6.2.0
gdown>=5.1.0
pre-commit>=3.7.0
pyarrow>=16.0.0
pyav>=12.0.5
tensorboard>=2.16.2
tqdm>=4.66.2
transformers>=4.36.2
rotary_embedding_torch>=0.5.3
pandarallel>=1.6.5
colorama>=0.4.6
backoff>=2.2.1
tiktoken>=0.6.0
openai>=1.27.0
wandb>=0.17.2
pynvml>=11.5.0
deepspeed>=0.14.4
vit_pytorch>=1.7.0
gym>=0.26.2